fixes #22 by setting the total paid
The issue was that the total due was still set to the amount already
invoiced.  At that point when the order was shipped it did a check
somewhere and the math didn't add up and it thought there was still
money owed.  Makes sense.  We used the Invoiced amount just incase as
well.